requirements.txt
streamlit>=1.37,<2
pandas>=2.0.0
numpy
yfinance
requests
isyatirimhisse
ta
scikit-learn
plotly
git+https://github.com/rongardF/tvdatafeed.git
openai
pytz
xgboost
optuna
shap
python-dotenv
pyarrow

aiohttp
